Refactor asp.net core test apps (#3558)
