+++ test/TestApp.AspNetCore/Controllers/ValuesController.cs
@@ -16,9 +16,7 @@
 using System.Collections.Generic;
 using Microsoft.AspNetCore.Mvc;
 
-#pragma warning disable SA1300 // Element should begin with upper-case letter
-namespace TestApp.AspNetCore._7._0.Controllers
-#pragma warning restore SA1300 // Element should begin with upper-case letter
+namespace TestApp.AspNetCore.Controllers
 {
     [Route("api/[controller]")]
     public class ValuesController : Controller
